Move classifier into models folder. Rename helical.py to base_models.py Move BaseTaskModel to base_models.py Move BaseModelProtocol to base_models.py
